requirements.txt
Django~=5.2
django-admin-sortable2==2.3.1
django-tinymce==5.0.0
environs==5.2.1